Fix NPE in onCreate method of LoaderCallbacks.
The NPE occurs when the rootinfo object's member variables are accessed
before it has been initialized (in other words when it is null). To get
around this issue, a null check is put on the rootinfo object and if it
is null, the stack is reset to the "Recents" root for the selected
userId.

Test: Manual
Bug: 347114093
Flag: EXEMPT bugfix
Change-Id: I1fd3ac55e605bb4e42494a3d4c9ecdb4b9765646